requirements.txt
certifi==2016.2.28
Django==2.0.7
django-autocomplete-light==3.3.0rc3
django-crispy-forms==1.7.2
django-extensions==2.1.0
django-filter==2.0.0
django-rest-framework==0.1.0
django-reversion==3.0.0
django-tables2==1.21.2
djangorestframework==3.8.2
numpy==1.14.5
pandas==0.23.3
psycopg2==2.7.5
python-dateutil==2.7.3
pytz==2018.5
scikit-learn==0.19.2
six==1.11.0
wincertstore==0.2
